# api/openapi.yaml
openapi: 3.1.0
info:
  title: Kredit API
  version: 0.1.0
  description: Contract-first API for Kredit production v1.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /webhooks/meta:
    get:
      security: []
      operationId: verifyMetaWebhookSubscription
      summary: Meta subscription challenge using the configured verification token
      responses:
        '200': {description: Plain-text challenge}
        '403': {description: Verification token did not match}
    post:
      security: []
      operationId: receiveMetaWebhook
      summary: Receive signed WhatsApp statuses and text messages
      description: X-Hub-Signature-256 authenticates the raw body. Delivery statuses must match the original encrypted route, phone account and recipient. Inbound replies do not execute financial commands.
      responses:
        '200': {description: Authenticated event handled}
        '403': {description: Signature invalid}
        '503': {description: Evidence cannot yet be linked or processing unavailable; retry required}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: kredit_session
...